# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Gestion Magasin Documentation",
        "description": "Implementation solution de Gestion de Magasin avec Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/receptions": {
            "get": {
                "tags": [
                    "Affiche Receptions et Lignereceptions"
                ],
                "summary": "Affiche la liste des Receptions et Lignereceptions",
                "description": "Retourner Touts les Receptions et Lignereceptions",
                "operationId": "Affiche Receptions et Lignereceptions",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Ajouter Receptions et Lignereceptions"
                ],
                "summary": "Ajoute Receptions et Lignereceptions",
                "description": "Ajouter  Receptions et Lignereceptions et Retourne les Valeurs Ajouter",
                "operationId": "Store Receptions et Lignereceptions",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/receptions/{reception}": {
            "get": {
                "tags": [
                    "Affiche Reception et Lignereception par ID"
                ],
                "summary": "Affiche le Commandeachat et Lignecommandeachat par ID",
                "description": "Retourner le Reception et Lignereception",
                "operationId": "Show Reception et Lignereception",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Mise en Jour Reception et Lignereception"
                ],
                "summary": "Modifier Reception et Lignereception",
                "description": "Retourner Reception et Lignereception Modifier",
                "operationId": "Modifier Reception et Lignereception",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Supprimer Reception et Lignereception"
                ],
                "summary": "Supprimer un Reception et Lignereception",
                "description": "Supprimer un reception et Lignereception",
                "operationId": "Delete Reception et Lignereception",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
...
    }
}